Refactor + Symbols support
This is a large commit that refactors a number of aspects in our
backend. Specifically, it improves in these areas:

 * Decouple the `ts_helpers` c functions into separate modules;
 * introduce an `adapt/` namespace, which contains all adapters
   to and from java and qdb;
 * standardize naming conventions;
 * migrate the "points" api into JNI;
 * lots of other quality-of-life improvements.

+++ src/main/c++/net/quasardb/qdb/jni/export/qdb_batch.cpp
@@ -1,69 +1,56 @@
-#include <qdb/batch.h>
-
-#include "net_quasardb_qdb_jni_qdb.h"
-
 #include "../env.h"
 #include "../exception.h"
 #include "../util/helpers.h"
+#include "net_quasardb_qdb_jni_qdb.h"
+#include <qdb/batch.h>
 
 namespace jni = qdb::jni;
 
-static qdb_operation_t &
-get_operation(jlong batch, jint index)
+static qdb_operation_t & get_operation(jlong batch, jint index)
 {
     return reinterpret_cast<qdb_operation_t *>(batch)[index];
 }
 
-extern "C" JNIEXPORT jint JNICALL
-Java_net_quasardb_qdb_jni_qdb_init_1operations(JNIEnv *jniEnv,
-                                               jclass /*thisClass*/,
-                                               jlong handle,
-                                               jint count,
-                                               jobject batch)
+extern "C" JNIEXPORT jint JNICALL Java_net_quasardb_qdb_jni_qdb_init_1operations(
+    JNIEnv * jniEnv, jclass /*thisClass*/, jlong handle, jint count, jobject batch)
 {
     qdb::jni::env env(jniEnv);
     try
     {
-        qdb_operation_t *ops = new qdb_operation_t[count];
-        jni::exception::throw_if_error((qdb_handle_t)handle,
-                                       {qdb_e_alias_not_found},
-                                       qdb_init_operations(ops, count));
+        qdb_operation_t * ops = new qdb_operation_t[count];
+        jni::exception::throw_if_error(
+            (qdb_handle_t)handle, {qdb_e_alias_not_found}, qdb_init_operations(ops, count));
         setLong(env, batch, reinterpret_cast<jlong>(ops));
         return qdb_e_ok;
     }
-    catch (jni::exception const &e)
+    catch (jni::exception const & e)
     {
         e.throw_new(env);
         return e.error();
     }
 }
 
-extern "C" JNIEXPORT jint JNICALL
-Java_net_quasardb_qdb_jni_qdb_delete_1batch(JNIEnv *jniEnv,
-                                            jclass /*thisClass*/,
-                                            jlong handle,
-                                            jlong batch)
+extern "C" JNIEXPORT jint JNICALL Java_net_quasardb_qdb_jni_qdb_delete_1batch(
+    JNIEnv * jniEnv, jclass /*thisClass*/, jlong handle, jlong batch)
 {
     qdb::jni::env env(jniEnv);
-    void *ptr = reinterpret_cast<void *>(batch);
+    void * ptr = reinterpret_cast<void *>(batch);
     qdb_release((qdb_handle_t)handle, ptr);
     return qdb_e_ok;
 }
 
-extern "C" JNIEXPORT jint JNICALL
-Java_net_quasardb_qdb_jni_qdb_run_1batch(
-    JNIEnv *jniEnv, jclass /*thisClass*/, jlong handle, jlong batch, jint count)
+extern "C" JNIEXPORT jint JNICALL Java_net_quasardb_qdb_jni_qdb_run_1batch(
+    JNIEnv * jniEnv, jclass /*thisClass*/, jlong handle, jlong batch, jint count)
 {
-
     qdb::jni::env env(jniEnv);
     try
     {
-        qdb_operation_t *ops = reinterpret_cast<qdb_operation_t *>(batch);
+        qdb_operation_t * ops = reinterpret_cast<qdb_operation_t *>(batch);
 
         // returns size_t, not qdb_error_t!
-        return qdb_run_batch((qdb_handle_t)handle, ops, count);
+        return (jint)qdb_run_batch((qdb_handle_t)handle, ops, count);
     }
-    catch (jni::exception const &e)
+    catch (jni::exception const & e)
     {
         e.throw_new(env);
         return e.error();
@@ -75,8 +62,7 @@ Java_net_quasardb_qdb_jni_qdb_run_1batch(
 // -----------------------
 
 extern "C" JNIEXPORT void JNICALL
-Java_net_quasardb_qdb_jni_qdb_batch_1write_1blob_1compare_1and_1swap(
-    JNIEnv *jniEnv,
+Java_net_quasardb_qdb_jni_qdb_batch_1write_1blob_1compare_1and_1swap(JNIEnv * jniEnv,
     jclass /*thisClass*/,
     jlong batch,
     jint index,
@@ -87,21 +73,18 @@ Java_net_quasardb_qdb_jni_qdb_batch_1write_1blob_1compare_1and_1swap(
 {
     qdb::jni::env env(jniEnv);
 
-    qdb_operation_t &op = get_operation(batch, index);
-    op.type = qdb_op_blob_cas;
-    op.alias = alias ? env.instance().GetStringUTFChars(alias, NULL) : NULL;
-    op.blob_cas.new_content = env.instance().GetDirectBufferAddress(newContent);
-    op.blob_cas.new_content_size =
-        (qdb_size_t)env.instance().GetDirectBufferCapacity(newContent);
-    op.blob_cas.comparand = env.instance().GetDirectBufferAddress(comparand);
-    op.blob_cas.comparand_size =
-        (qdb_size_t)env.instance().GetDirectBufferCapacity(comparand);
-    op.blob_cas.expiry_time = expiry;
+    qdb_operation_t & op         = get_operation(batch, index);
+    op.type                      = qdb_op_blob_cas;
+    op.alias                     = alias ? env.instance().GetStringUTFChars(alias, NULL) : NULL;
+    op.blob_cas.new_content      = env.instance().GetDirectBufferAddress(newContent);
+    op.blob_cas.new_content_size = (qdb_size_t)env.instance().GetDirectBufferCapacity(newContent);
+    op.blob_cas.comparand        = env.instance().GetDirectBufferAddress(comparand);
+    op.blob_cas.comparand_size   = (qdb_size_t)env.instance().GetDirectBufferCapacity(comparand);
+    op.blob_cas.expiry_time      = expiry;
 }
 
 extern "C" JNIEXPORT jint JNICALL
-Java_net_quasardb_qdb_jni_qdb_batch_1read_1blob_1compare_1and_1swap(
-    JNIEnv *jniEnv,
+Java_net_quasardb_qdb_jni_qdb_batch_1read_1blob_1compare_1and_1swap(JNIEnv * jniEnv,
     jclass /*thisClass*/,
     jlong handle,
     jlong batch,
@@ -113,16 +96,15 @@ Java_net_quasardb_qdb_jni_qdb_batch_1read_1blob_1compare_1and_1swap(
 
     try
     {
-
-        qdb_operation_t &op = get_operation(batch, index);
+        qdb_operation_t & op = get_operation(batch, index);
         jni::exception::throw_if_error((qdb_handle_t)handle, op.error);
 
         if (alias) env.instance().ReleaseStringUTFChars(alias, op.alias);
-        setByteBuffer(env, originalContent, op.blob_cas.original_content,
-                      op.blob_cas.original_content_size);
+        setByteBuffer(
+            env, originalContent, op.blob_cas.original_content, op.blob_cas.original_content_size);
         return op.error;
     }
-    catch (jni::exception const &e)
+    catch (jni::exception const & e)
     {
         e.throw_new(env);
         return e.error();
@@ -133,42 +115,37 @@ Java_net_quasardb_qdb_jni_qdb_batch_1read_1blob_1compare_1and_1swap(
 // blob_get
 // -----------------------
 
-extern "C" JNIEXPORT void JNICALL
-Java_net_quasardb_qdb_jni_qdb_batch_1write_1blob_1get(JNIEnv *jniEnv,
-                                                      jclass /*thisClass*/,
-                                                      jlong batch,
-                                                      jint index,
-                                                      jstring alias)
+extern "C" JNIEXPORT void JNICALL Java_net_quasardb_qdb_jni_qdb_batch_1write_1blob_1get(
+    JNIEnv * jniEnv, jclass /*thisClass*/, jlong batch, jint index, jstring alias)
 {
     qdb::jni::env env(jniEnv);
 
-    qdb_operation_t &op = get_operation(batch, index);
-    op.type = qdb_op_blob_get;
-    op.alias = alias ? env.instance().GetStringUTFChars(alias, NULL) : NULL;
+    qdb_operation_t & op = get_operation(batch, index);
+    op.type              = qdb_op_blob_get;
+    op.alias             = alias ? env.instance().GetStringUTFChars(alias, NULL) : NULL;
 }
 
-extern "C" JNIEXPORT jint JNICALL
-Java_net_quasardb_qdb_jni_qdb_batch_1read_1blob_1get(JNIEnv *jniEnv,
-                                                     jclass /*thisClass*/,
-                                                     jlong handle,
-                                                     jlong batch,
-                                                     jint index,
-                                                     jstring alias,
-                                                     jobject content)
+extern "C" JNIEXPORT jint JNICALL Java_net_quasardb_qdb_jni_qdb_batch_1read_1blob_1get(
+    JNIEnv * jniEnv,
+    jclass /*thisClass*/,
+    jlong handle,
+    jlong batch,
+    jint index,
+    jstring alias,
+    jobject content)
 {
     qdb::jni::env env(jniEnv);
 
     try
     {
-        qdb_operation_t &op = get_operation(batch, index);
+        qdb_operation_t & op = get_operation(batch, index);
         jni::exception::throw_if_error((qdb_handle_t)handle, op.error);
 
         if (alias) env.instance().ReleaseStringUTFChars(alias, op.alias);
-        setByteBuffer(env, content, op.blob_get.content,
-                      op.blob_get.content_size);
+        setByteBuffer(env, content, op.blob_get.content, op.blob_get.content_size);
         return op.error;
     }
-    catch (jni::exception const &e)
+    catch (jni::exception const & e)
     {
         e.throw_new(env);
         return e.error();
@@ -180,8 +157,7 @@ Java_net_quasardb_qdb_jni_qdb_batch_1read_1blob_1get(JNIEnv *jniEnv,
 // -----------------------
 
 extern "C" JNIEXPORT void JNICALL
-Java_net_quasardb_qdb_jni_qdb_batch_1write_1blob_1get_1and_1update(
-    JNIEnv *jniEnv,
+Java_net_quasardb_qdb_jni_qdb_batch_1write_1blob_1get_1and_1update(JNIEnv * jniEnv,
     jclass /*thisClass*/,
     jlong batch,
     jint index,
@@ -191,19 +167,17 @@ Java_net_quasardb_qdb_jni_qdb_batch_1write_1blob_1get_1and_1update(
 {
     qdb::jni::env env(jniEnv);
 
-    qdb_operation_t &op = get_operation(batch, index);
-    op.type = qdb_op_blob_get_and_update;
-    op.alias = alias ? env.instance().GetStringUTFChars(alias, NULL) : NULL;
-    op.blob_get_and_update.new_content =
-        env.instance().GetDirectBufferAddress(content);
+    qdb_operation_t & op = get_operation(batch, index);
+    op.type              = qdb_op_blob_get_and_update;
+    op.alias             = alias ? env.instance().GetStringUTFChars(alias, NULL) : NULL;
+    op.blob_get_and_update.new_content = env.instance().GetDirectBufferAddress(content);
     op.blob_get_and_update.new_content_size =
         (qdb_size_t)env.instance().GetDirectBufferCapacity(content);
     op.blob_get_and_update.expiry_time = expiry;
 }
 
-extern "C" JNIEXPORT jint JNICALL
-Java_net_quasardb_qdb_jni_qdb_batch_1read_1blob_1get_1and_1update(
-    JNIEnv *jniEnv,
+extern "C" JNIEXPORT jint JNICALL Java_net_quasardb_qdb_jni_qdb_batch_1read_1blob_1get_1and_1update(
+    JNIEnv * jniEnv,
     jclass /*thisClass*/,
     jlong handle,
     jlong batch,
@@ -214,15 +188,15 @@ Java_net_quasardb_qdb_jni_qdb_batch_1read_1blob_1get_1and_1update(
     qdb::jni::env env(jniEnv);
     try
     {
-        qdb_operation_t &op = get_operation(batch, index);
+        qdb_operation_t & op = get_operation(batch, index);
         jni::exception::throw_if_error((qdb_handle_t)handle, op.error);
 
         if (alias) env.instance().ReleaseStringUTFChars(alias, op.alias);
         setByteBuffer(env, content, op.blob_get_and_update.original_content,
-                      op.blob_get_and_update.original_content_size);
+            op.blob_get_and_update.original_content_size);
         return op.error;
     }
-    catch (jni::exception const &e)
+    catch (jni::exception const & e)
     {
         e.throw_new(env);
         return e.error();
@@ -233,46 +207,39 @@ Java_net_quasardb_qdb_jni_qdb_batch_1read_1blob_1get_1and_1update(
 // blob_put
 // -----------------------
 
-extern "C" JNIEXPORT void JNICALL
-Java_net_quasardb_qdb_jni_qdb_batch_1write_1blob_1put(JNIEnv *jniEnv,
-                                                      jclass /*thisClass*/,
-                                                      jlong batch,
-                                                      jint index,
-                                                      jstring alias,
-                                                      jobject content,
-                                                      jlong expiry)
+extern "C" JNIEXPORT void JNICALL Java_net_quasardb_qdb_jni_qdb_batch_1write_1blob_1put(
+    JNIEnv * jniEnv,
+    jclass /*thisClass*/,
+    jlong batch,
+    jint index,
+    jstring alias,
+    jobject content,
+    jlong expiry)
 {
     qdb::jni::env env(jniEnv);
 
-    qdb_operation_t &op = get_operation(batch, index);
-    op.type = qdb_op_blob_put;
-    op.alias = alias ? env.instance().GetStringUTFChars(alias, NULL) : NULL;
-    op.blob_put.content = env.instance().GetDirectBufferAddress(content);
-    op.blob_put.content_size =
-        (qdb_size_t)env.instance().GetDirectBufferCapacity(content);
-    op.blob_put.expiry_time = expiry;
+    qdb_operation_t & op     = get_operation(batch, index);
+    op.type                  = qdb_op_blob_put;
+    op.alias                 = alias ? env.instance().GetStringUTFChars(alias, NULL) : NULL;
+    op.blob_put.content      = env.instance().GetDirectBufferAddress(content);
+    op.blob_put.content_size = (qdb_size_t)env.instance().GetDirectBufferCapacity(content);
+    op.blob_put.expiry_time  = expiry;
 }
 
-extern "C" JNIEXPORT jint JNICALL
-Java_net_quasardb_qdb_jni_qdb_batch_1read_1blob_1put(JNIEnv *jniEnv,
-                                                     jclass /*thisClass*/,
-                                                     jlong handle,
-                                                     jlong batch,
-                                                     jint index,
-                                                     jstring alias)
+extern "C" JNIEXPORT jint JNICALL Java_net_quasardb_qdb_jni_qdb_batch_1read_1blob_1put(
+    JNIEnv * jniEnv, jclass /*thisClass*/, jlong handle, jlong batch, jint index, jstring alias)
 {
     qdb::jni::env env(jniEnv);
 
     try
     {
-
-        qdb_operation_t &op = get_operation(batch, index);
+        qdb_operation_t & op = get_operation(batch, index);
         jni::exception::throw_if_error((qdb_handle_t)handle, op.error);
 
         if (alias) env.instance().ReleaseStringUTFChars(alias, op.alias);
         return op.error;
     }
-    catch (jni::exception const &e)
+    catch (jni::exception const & e)
     {
         e.throw_new(env);
         return e.error();
@@ -283,45 +250,39 @@ Java_net_quasardb_qdb_jni_qdb_batch_1read_1blob_1put(JNIEnv *jniEnv,
 // blob_update
 // -----------------------
 
-extern "C" JNIEXPORT void JNICALL
-Java_net_quasardb_qdb_jni_qdb_batch_1write_1blob_1update(JNIEnv *jniEnv,
-                                                         jclass /*thisClass*/,
-                                                         jlong batch,
-                                                         jint index,
-                                                         jstring alias,
-                                                         jobject content,
-                                                         jlong expiry)
+extern "C" JNIEXPORT void JNICALL Java_net_quasardb_qdb_jni_qdb_batch_1write_1blob_1update(
+    JNIEnv * jniEnv,
+    jclass /*thisClass*/,
+    jlong batch,
+    jint index,
+    jstring alias,
+    jobject content,
+    jlong expiry)
 {
     qdb::jni::env env(jniEnv);
 
-    qdb_operation_t &op = get_operation(batch, index);
-    op.type = qdb_op_blob_update;
-    op.alias = alias ? env.instance().GetStringUTFChars(alias, NULL) : NULL;
-    op.blob_put.content = env.instance().GetDirectBufferAddress(content);
-    op.blob_put.content_size =
-        (qdb_size_t)env.instance().GetDirectBufferCapacity(content);
-    op.blob_put.expiry_time = expiry;
+    qdb_operation_t & op     = get_operation(batch, index);
+    op.type                  = qdb_op_blob_update;
+    op.alias                 = alias ? env.instance().GetStringUTFChars(alias, NULL) : NULL;
+    op.blob_put.content      = env.instance().GetDirectBufferAddress(content);
+    op.blob_put.content_size = (qdb_size_t)env.instance().GetDirectBufferCapacity(content);
+    op.blob_put.expiry_time  = expiry;
 }
 
-extern "C" JNIEXPORT jint JNICALL
-Java_net_quasardb_qdb_jni_qdb_batch_1read_1blob_1update(JNIEnv *jniEnv,
-                                                        jclass /*thisClass*/,
-                                                        jlong handle,
-                                                        jlong batch,
-                                                        jint index,
-                                                        jstring alias)
+extern "C" JNIEXPORT jint JNICALL Java_net_quasardb_qdb_jni_qdb_batch_1read_1blob_1update(
+    JNIEnv * jniEnv, jclass /*thisClass*/, jlong handle, jlong batch, jint index, jstring alias)
 {
     qdb::jni::env env(jniEnv);
 
     try
     {
-        qdb_operation_t &op = get_operation(batch, index);
+        qdb_operation_t & op = get_operation(batch, index);
         jni::exception::throw_if_error((qdb_handle_t)handle, op.error);
 
         if (alias) env.instance().ReleaseStringUTFChars(alias, op.alias);
         return op.error;
     }
-    catch (jni::exception const &e)
+    catch (jni::exception const & e)
     {
         e.throw_new(env);
         return e.error();
